requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
telebot==0.0.5
tweepy==4.14.0
urllib3==2.0.7
Werkzeug==3.0.0
